Update .gitignore and refactor news modules for improved structure and functionality
