+++ PythonProject3/main.py
@@ -1,14 +1,11 @@
 from __future__ import annotations
-
 from datetime import datetime
-
 import requests
-
-from ApexNews import ApexNews, get_existing_entries as get_apex_existing_entries
+from video_games.ApexNews import ApexNews, get_existing_entries as get_apex_existing_entries
 from Discord import SendToDiscord
-from GamingNews import ArcRaidersNews, get_existing_entries as get_game_existing_entries
-from HackingNews import NewsFeed, get_existing_entries
-from LeagueNews import LeagueNews, get_existing_entries as get_league_existing_entries
+from video_games.GamingNews import ArcRaidersNews, get_existing_entries as get_game_existing_entries
+from news.HackingNews import NewsFeed, get_existing_entries
+from video_games.LeagueNews import LeagueNews, get_existing_entries as get_league_existing_entries
 from webhook import webhook
 
 
@@ -48,14 +45,14 @@ def main():
     except Exception as e:
         print(f"Failed to fetch game news: {e}")
 
-    game_existing_entries = get_game_existing_entries()
+    game_existing_entries = get_game_existing_entries('arc_raiders_news.txt')
     print(f"Existing game entries: {game_existing_entries}")
 
     for article in game_news.news:
         try:
             article_date = datetime.strptime(article['date'], '%B %d, %Y').date()
         except Exception as e:
-            print(f"Could not parse date for article: {article}")
+            print(f"Could not parse date for article: {e}")
             continue
         article_key = article['link']
         print(f"Checking game entry: (Title: {article['title']}, Date: {article_date}, Link: {article_key})")
@@ -74,14 +71,14 @@ def main():
     except Exception as e:
         print(f"Failed to fetch League news: {e}")
 
-    league_existing_entries = get_league_existing_entries()
+    league_existing_entries = get_league_existing_entries('league_news.txt')
     print(f"Existing League entries: {league_existing_entries}")
 
     for article in league_news.news:
         try:
             article_date = datetime.strptime(article['date'], '%B %d, %Y').date()
         except Exception as e:
-            print(f"Could not parse date for article: {article}")
+            print(f"Could not parse date for article: {e}")
             continue
         article_key = article['link']
         print(f"Checking League entry: (Title: {article['title']}, Date: {article_date}, Link: {article_key})")
@@ -100,14 +97,14 @@ def main():
     except Exception as e:
         print(f"Failed to fetch Apex Legends news: {e}")
 
-    apex_existing_entries = get_apex_existing_entries()
+    apex_existing_entries = get_apex_existing_entries('apex_news.txt')
     print(f"Existing Apex entries: {apex_existing_entries}")
 
     for article in apex_news.news:
         try:
             article_date = datetime.strptime(article['date'], '%B %d, %Y').date()
         except Exception as e:
-            print(f"Invalid or missing date format for Apex article: {article}")
+            print(f"Invalid or missing date format for Apex article: {e}")
             continue
         article_key = article['link']
         print(f"Checking Apex entry: (Title: {article['title']}, Date: {article_date}, Link: {article_key})")
