+++ PythonProject3/requirements.txt
@@ -1,4 +1,4 @@
-requests~=2.34.1
+requests~=2.34.2
 discord.py
 feedparser~=6.0.11
 python-dotenv~=1.2.1
